ai-pokemon-image/requirements.txt
rembg
flask
flask-cors
pillow
onnxruntime

